requirements.txt
crewai
crewai_tools
langchain_community
mammoth
markdown
python-dotenv
gradio